fix(pisa-proxy, runtime): reafctor server complete
