+++ pisa-proxy/protocol/mysql/src/server/mod.rs
@@ -12,9 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod conn;
 pub mod err;
-pub mod packet;
 pub mod stream;
 mod tls;
 pub mod codec;
